# swagger.yaml
openapi: 3.0.1
info:
  title: Get Science API
  description: API платформы для научных и образовательных мероприятий
  version: 1.0.0
  contact:
    name: Get Science Team
    url: https://get-science-events.ru
servers:
  - url: https://get-science-events.ru/api/v1
    description: Основной сервер
...
paths:
...
  /applications/event/{eventId}:
    get:
      tags:
        - applications
      summary: Получение списка заявок для мероприятия
      description: Получает список всех заявок для указанного мероприятия
      operationId: getApplicationsByEvent
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список заявок успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationResponse'
        '404':
          description: Мероприятие не найдено
...
  /events:
    post:
      tags:
        - events
      summary: Создание нового мероприятия
      description: Создает новое мероприятие с возможностью загрузки файлов
      operationId: createEventMultipart
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                eventRequest:
                  $ref: '#/components/schemas/EventRequest'
                fileEventRequest:
                  type: array
                  items:
                    $ref: '#/components/schemas/FileEventRequest'
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Мероприятие успешно создано
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Ошибка валидации данных
    get:
      tags:
        - events
      summary: Получение списка всех мероприятий
      description: Получает список всех мероприятий с пагинацией
      operationId: getAllEventsPaged
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 12
      responses:
        '200':
          description: Список мероприятий успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageableResponseEventResponse'
  /events/{eventId}:
    put:
      tags:
        - events
      summary: Обновление мероприятия
      description: Обновляет существующее мероприятие
      operationId: updateEventMultipart
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                eventRequest:
                  $ref: '#/components/schemas/EventRequest'
                fileEventRequest:
                  type: array
                  items:
                    $ref: '#/components/schemas/FileEventRequest'
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: Мероприятие успешно обновлено
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Ошибка валидации данных
        '404':
          description: Мероприятие не найдено
    get:
      tags:
        - events
      summary: Получение мероприятия по идентификатору
      description: Получает информацию о мероприятии по его идентификатору
      operationId: getEventById
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о мероприятии успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '404':
          description: Мероприятие не найдено
    delete:
      tags:
        - events
      summary: Удаление мероприятия
      description: Удаляет мероприятие по его идентификатору
      operationId: deleteEvent
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Мероприятие успешно удалено
          content:
            application/json:
              schema:
                type: string
        '400':
          description: Ошибка валидации данных
        '404':
          description: Мероприятие не найдено
  /events/{eventId}/recommendations:
    get:
      tags:
        - events
      summary: Получение рекомендаций мероприятий
      description: Получает список рекомендуемых мероприятий на основе указанного мероприятия
      operationId: searchEvents
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список рекомендаций успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventResponse'
        '404':
          description: Мероприятие не найдено
  /events/{eventId}/files:
    post:
      tags:
        - events
      summary: Добавление файла к мероприятию
      description: Добавляет файл к указанному мероприятию
      operationId: addFileToEvent
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileEventRequest'
      responses:
        '200':
          description: Файл успешно добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileEventResponse'
        '400':
          description: Ошибка валидации данных
        '404':
          description: Мероприятие не найдено
    get:
      tags:
        - events
      summary: Получение файлов мероприятия
      description: Получает список файлов, прикрепленных к мероприятию
      operationId: getFilesByEvent
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список файлов успешно получен
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileEventResponse'
        '404':
          description: Мероприятие не найдено
  /events/files/{fileId}:
    get:
      tags:
        - events
      summary: Получение файла мероприятия по идентификатору
      description: Получает информацию о файле мероприятия по его идентификатору
      operationId: getFileEventById
      parameters:
        - name: fileId
          in: path
          description: Идентификатор файла
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация о файле успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileEventResponse'
        '404':
          description: Файл не найден
  /events/my-events:
    get:
      tags:
        - events
      summary: Получение списка мероприятий текущего пользователя
      description: Получает список мероприятий, созданных текущим пользователем, с пагинацией
      operationId: getEventsByJwtPaged
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 12
      responses:
        '200':
          description: Список мероприятий успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageableResponseEventResponse'
        '401':
          description: Не авторизован
  /events/search:
    post:
      tags:
        - events
      summary: Расширенный поиск мероприятий
      description: Поиск мероприятий с поддержкой множественных фильтров
      operationId: searchEventsWithMultipleFilters
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventFilterRequest'
      responses:
        '200':
          description: Результаты поиска успешно получены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageableResponseEventResponse'
  /events/organizer/{organizerId}:
    get:
      tags:
        - events
      summary: Получение списка мероприятий организатора
      description: Получает список мероприятий, созданных указанным организатором, с пагинацией
      operationId: getEventsByOrganizerPaged
      parameters:
        - name: organizerId
          in: path
          description: Идентификатор организатора
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          description: Номер страницы (начиная с 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Размер страницы
          required: false
          schema:
            type: integer
            default: 12
      responses:
        '200':
          description: Список мероприятий успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageableResponseEventResponse'
        '404':
          description: Организатор не найден
  /events/{eventId}/with-people:
    get:
      tags:
        - events
      summary: Получение мероприятия с информацией о рецензентах и соорганизаторах
      description: Получает информацию о мероприятии вместе с данными о рецензентах и соорганизаторах
      operationId: getEventWithReviewersAndCoowners
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информация успешно получена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '404':
          description: Мероприятие не найдено
...
  /events/{eventId}/calendar:
    get:
      tags:
        - events
      summary: Экспорт мероприятия в формате iCalendar
      description: Экспортирует мероприятие в формате iCalendar (.ics)
      operationId: exportEventCalendar
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Файл iCalendar успешно сгенерирован
          content:
            text/calendar:
              schema:
                type: string
                format: binary
        '404':
          description: Мероприятие не найдено
  /events/{eventId}/participants/export:
    get:
      tags:
        - events
      summary: Экспорт списка участников в Excel
      description: Экспортирует список участников мероприятия в Excel
      operationId: exportEventParticipantsExcel
      parameters:
        - name: eventId
          in: path
          description: Идентификатор мероприятия
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Файл Excel успешно сгенерирован
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401':
          description: Не авторизован
        '404':
          description: Мероприятие не найдено
...
components:
  schemas:
...
    ApplicationResponse:
      type: object
      properties:
        applicationId:
          type: integer
          format: int64
          description: Идентификатор заявки
        eventId:
          type: integer
          format: int64
          description: Идентификатор мероприятия
        eventName:
          type: string
          description: Название мероприятия
        profileId:
          type: integer
          format: int64
          description: Идентификатор профиля
        status:
          type: string
          description: Статус заявки
        submissionDate:
          type: string
          format: date-time
          description: Дата подачи заявки
        message:
          type: string
          description: Сообщение от заявителя
          nullable: true
        isObserver:
          type: boolean
          description: Является ли наблюдателем
        verdict:
          type: string
          description: Решение по заявке
          nullable: true
        fileApplications:
          type: array
          description: Файлы заявки
          items:
            $ref: '#/components/schemas/FileApplicationResponse'
...
    FileEventRequest:
      type: object
      properties:
        fileName:
          type: string
          description: Название файла
        filePath:
          type: string
          description: Путь к файлу
        uploadDate:
          type: string
          format: date-time
          description: Дата загрузки
        fileType:
          type: string
          description: Тип файла
        fileCategory:
          type: string
          description: Категория файла
        description:
          type: string
          description: Описание файла
          nullable: true
    FileEventResponse:
      type: object
      properties:
        fileId:
          type: integer
          format: int64
          description: Идентификатор файла
        fileName:
          type: string
          description: Название файла
        filePath:
          type: string
          description: Путь к файлу
        uploadDate:
          type: string
          format: date-time
          description: Дата загрузки
        fileType:
          type: string
          description: Тип файла
        fileCategory:
          type: string
          description: Категория файла
        description:
          type: string
          description: Описание файла
          nullable: true
        downloadUrl:
          type: string
          description: URL для скачивания
          nullable: true
...
    PageableResponseEventResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EventResponse'
        hasNext:
          type: boolean
          description: Есть ли следующая страница
        totalPages:
          type: integer
          description: Общее количество страниц
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []